Completed Task system
Added the ability to pickup and complete tasks both in a good and bad way, with points awarded accordingly.

Added the ability to create a task and assign it steps aswell as assign the steps specific rooms where the step is completed upon entry(for now).

Made so the parser now can take 3 words, so it is possible to show task specific information aswell as a complete overview of active tasks.